requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.3.0
python-dateutil==2.9.0
django-debug-toolbar==4.3.0
pytest-django==4.8.0
pytest==8.1.1
factory-boy==3.3.0django-htmx==1.17.0
